refactor: move schema to migrations folder
