+++ project/lib/db/queries/comments.ts
@@ -1,4 +1,4 @@
-import { comments } from "@/lib/db/drizzle/schema";
+import { comments } from "@/lib/db/drizzle/migrations/schema";
 import {
   Comment,
   CreateCommentPayload,
