+++ project/types/comments.d.ts
@@ -1,4 +1,4 @@
-import { comments } from "@/lib/db/drizzle/schema";
+import { comments } from "@/lib/db/drizzle/migrations/schema";
 import { InferSelectModel } from "drizzle-orm";
 
 export type Comment = InferSelectModel<typeof comments>;
